some fixes in update system

+++ lib/core/lhcore/lhupdate.php
@@ -7,28 +7,44 @@ class erLhcoreClassUpdate {
         return (self::VERSION)/100;
     }
 		
-	public static function doTablesUpdate($definition){
-		$updateInformation = self::getTablesStatus($definition);
-		$db = ezcDbInstance::get();
+    public static function doTablesUpdate($definition, $executeQueries = true) {
 		
 		$errorMessages = array();
+		$queries = array();
 		
+		$updateInformation = self::getTablesStatus($definition);
+		
+		$db = ezcDbInstance::get();
+			
 		foreach ($updateInformation as $table => $tableData) {
+			
 			if ($tableData['error'] == true) {
+				
 				foreach ($tableData['queries'] as $query) {
+					
 					try {
-						$db->query($query);
+						
+						if ($executeQueries) {
+							$db->query($query);
+						}
+						
+						$queries[] = $query;
+						
 					} catch (Exception $e) {
 						$errorMessages[] = $e->getMessage();
 					}
+					
 				}
+				
 			}
+			
 		}
 		
-		return $errorMessages;		
+		return array('error_messages' => $errorMessages, 'queries' => $queries);		
 	}
 	
-	public static function getTablesStatus($definition){
+	public static function getTablesStatus($definition) {
+		
 		$db = ezcDbInstance::get();
 		
 		$tablesStatus = array();		
@@ -43,6 +59,19 @@ class erLhcoreClassUpdate {
 				
 				$status = array();
 				
+				foreach ($columnsData as $columnExisting) {
+					$columnFound = false;				
+					foreach ($columnsDesired as $column) {
+						if ($columnExisting['field'] == $column['field']) {
+							$columnFound = true;							
+						}
+					}
+					
+					if ($columnFound == false){
+						$status[] = "[{$columnExisting['field']}] column not defined in JSON file!";		
+					}			
+				}
+				
 				foreach ($columnsDesired as $columnDesired) {
 					$columnFound = false;
 					$typeMatch = true;
@@ -55,12 +84,12 @@ class erLhcoreClassUpdate {
 							}
 						}	
 					}
+
 					if ($typeMatch == false) {
 						$tablesStatus[$table]['error'] = true;
 						$status[] = "[{$columnDesired['field']}] column type is not correct";
 												
-						$tablesStatus[$table]['queries'][] = "ALTER TABLE `{$table}`
-						CHANGE `{$columnDesired['field']}` `{$columnDesired['field']}` {$columnDesired['type']} NOT NULL;";
+						$tablesStatus[$table]['queries'][] = "ALTER TABLE `{$table}` CHANGE `{$columnDesired['field']}` `{$columnDesired['field']}` {$columnDesired['type']} NOT NULL;";
 					}
 					
 					if ($columnFound == false) {
@@ -68,13 +97,11 @@ class erLhcoreClassUpdate {
 						$status[] = "[{$columnDesired['field']}] column was not found";
 						
 						$default = '';
-						if ($columnDesired['default'] != null){
+						if ($columnDesired['default'] !== null){
 							$default = " DEFAULT '{$columnDesired['default']}'";
 						}
 								
-						$tablesStatus[$table]['queries'][] = "ALTER TABLE `{$table}`
-						ADD `{$columnDesired['field']}` {$columnDesired['type']} NOT NULL{$default},
-						COMMENT='';";
+						$tablesStatus[$table]['queries'][] = "ALTER TABLE `{$table}` ADD `{$columnDesired['field']}` {$columnDesired['type']} NOT NULL{$default}, COMMENT='';";
 					}					
 				}
 				
@@ -90,44 +117,48 @@ class erLhcoreClassUpdate {
 			}			
 		}
 				
-		foreach ($definition['tables_indexes'] as $table => $dataTableIndex) {		    
-		    try {
-    		    $sql = 'SHOW INDEX FROM '.$table;
-    		    $stmt = $db->prepare($sql);
-    		    $stmt->execute();
-    		    $columnsData = $stmt->fetchAll(PDO::FETCH_ASSOC); 
-    		    $status = array();
-    		    
-    		    $existingIndexes = array();
-    		    foreach ($columnsData as $indexData) {
-    		        $existingIndexes[] = $indexData['key_name'];
-    		    }
-    		    
-    		    $existingIndexes = array_unique($existingIndexes);
-    		    
-    		    $newIndexes = array_diff(array_keys($dataTableIndex['new']), $existingIndexes);
-    		    
-    		    foreach ($newIndexes as $newIndex) {
-    		        $tablesStatus[$table]['queries'][] = $dataTableIndex['new'][$newIndex];
-    		        $status[] = "{$newIndex} index was not found";
-    		    }
-    		    
-    		    $removeIndexes = array_intersect($dataTableIndex['old'], $existingIndexes);
-    		   
-    		    foreach ($removeIndexes as $removeIndex) {
-    		        $tablesStatus[$table]['queries'][] = "ALTER TABLE `{$table}` DROP INDEX `{$removeIndex}`;";
-    		        $tablesStatus[$table]['error'] = true;
-    		        $status[] = "{$removeIndex} legacy index was found";
-    		    }
-    		    
-    		    if (!empty($status)) {
-    		        $tablesStatus[$table]['status'] = implode(", ", $status);
-    		        $tablesStatus[$table]['error'] = true;
-    		    }
-    		    
-		    } catch (Exception $e) {
-		        // Just not existing table perhaps
-		    }	    
+		if (isset($definition['tables_indexes'])) {
+			
+			foreach ($definition['tables_indexes'] as $table => $dataTableIndex) {
+				try {
+					$sql = 'SHOW INDEX FROM '.$table;
+					$stmt = $db->prepare($sql);
+					$stmt->execute();
+					$columnsData = $stmt->fetchAll(PDO::FETCH_ASSOC);
+					$status = array();
+			
+					$existingIndexes = array();
+					foreach ($columnsData as $indexData) {
+						$existingIndexes[] = $indexData['key_name'];
+					}
+			
+					$existingIndexes = array_unique($existingIndexes);
+			
+					$newIndexes = array_diff(array_keys($dataTableIndex['new']), $existingIndexes);
+			
+					foreach ($newIndexes as $newIndex) {
+						$tablesStatus[$table]['queries'][] = $dataTableIndex['new'][$newIndex];
+						$status[] = "{$newIndex} index was not found";
+					}
+			
+					$removeIndexes = array_intersect($dataTableIndex['old'], $existingIndexes);
+						
+					foreach ($removeIndexes as $removeIndex) {
+						$tablesStatus[$table]['queries'][] = "ALTER TABLE `{$table}` DROP INDEX `{$removeIndex}`;";
+						$tablesStatus[$table]['error'] = true;
+						$status[] = "{$removeIndex} legacy index was found";
+					}
+			
+					if (!empty($status)) {
+						$tablesStatus[$table]['status'] = implode(", ", $status);
+						$tablesStatus[$table]['error'] = true;
+					}
+			
+				} catch (Exception $e) {
+					// Just not existing table perhaps
+				}
+			}
+			
 		}
 				
 		foreach ($definition['tables_data'] as $table => $dataTable) {
@@ -136,6 +167,7 @@ class erLhcoreClassUpdate {
 			$status = array();
 			// Check that table has all required records
 			foreach ($dataTable as $record) {	
+
 				try {
 					$sql = "SELECT COUNT(*) as total_records FROM `{$table}` WHERE `{$tableIdentifier}` = :identifier_value";				
 					$stmt = $db->prepare($sql);
@@ -175,6 +207,155 @@ class erLhcoreClassUpdate {
 		}
 		
 		return $tablesStatus;
+	}
+		
+	public static function getElasticStatus($definition) {
+		
+		$typeStatus = array();		
+		
+		$cfg = erConfigClassLhConfig::getInstance();
+		$elasticIndex = $cfg->getSetting('elasticsearch','index');
+		
+		$elasticData = erLhcoreClassElasticClient::instance()->indices()->getMapping(array('index' => $elasticIndex));
+				
+		$currentMappingData = $elasticData[$elasticIndex]['mappings'];
+				
+		foreach ($definition['types'] as $type => $typeDefinition) {
+
+			if(isset($currentMappingData[$type])) {
+				
+				$status = array();
+				
+				$currentTypeProperties = $currentMappingData[$type]['properties'];
+				
+				// Add property
+				foreach ($typeDefinition as $property => $propertyData) {
+					
+					if(!isset($currentTypeProperties[$property])) {
+												
+						$status[] = '['.$property.'] property not found';
+						
+						$params = array(
+							'index' => $elasticIndex,
+							'type' => $type,
+							'body' => array(
+								$type => array(
+									'properties' => array($property => $propertyData)
+								)
+							)
+						);
+						
+						$typeStatus[$type]['actions']['type_property_add'][] = $params;
+						
+					}	
+					
+				}
+
+				// Remove property
+				/*				
+				foreach (array_keys($currentTypeProperties) as $property) {
+
+ 					if(!isset($typeDefinition[$property])) {
+						
+ 						$status[] = '['.$property.'] property removed';
+						
+ 						$typeStatus[$type]['actions']['type_property_delete'][] = array();
+						
+ 					}
+					
+				}
+				*/
+								
+				if (!empty($status)) {
+					$typeStatus[$type]['error'] = true;
+					$typeStatus[$type]['status'] = implode(', ', $status);
+				}				
+				
+			} else {
+				
+				// Add types
+				$typeStatus[$type]['error'] = true;
+				$typeStatus[$type]['status'] = 'type add';
+				
+				$params = array(
+					'index' => $elasticIndex,
+					'type' => $type,
+					'body' => array(
+						$type => array(
+							'properties' => $typeDefinition
+						)
+					)
+				);
+							
+				$typeStatus[$type]['actions']['type_add'][] = $params;
+				
+			}
+			
+		}	
+		
+		// Remove types		
+		foreach (array_keys($currentMappingData) as $type) {
+			
+ 			if(!isset($definition['types'][$type])) {
+ 				
+ 				$typeStatus[$type]['error'] = true;
+ 				$typeStatus[$type]['status'] = 'type removed';
+ 				
+ 				$params = array(
+ 					'index' => $elasticIndex,
+ 					'type' => $type
+ 				);			
+ 				
+ 				$typeStatus[$type]['actions']['type_delete'][] = $params;
+ 			
+ 			}		
+			
+		}
+		
+		return $typeStatus;
+		
+	}
+	
+	public static function doElasticUpdate($definition) {
+		
+		$errorMessages = array();
+		
+		$updateInformation = self::getElasticStatus($definition);
+		
+		foreach ($updateInformation as $type => $typeData) {
+			
+			if ($typeData['error'] == true) {
+				
+				foreach ($typeData['actions'] as $actionType => $actionParams) {
+					
+					foreach ($actionParams as $params) {
+						
+						try {
+								
+							if($actionType == 'type_add') {
+								erLhcoreClassElasticClient::instance()->indices()->putMapping($params);
+							} elseif($actionType == 'type_delete') {
+								erLhcoreClassElasticClient::instance()->indices()->deleteMapping($params);
+							} elseif($actionType == 'type_property_add') {
+								erLhcoreClassElasticClient::instance()->indices()->putMapping($params);
+							} elseif($actionType == 'type_property_delete') {
+								// Not used now
+							}
+						
+						} catch (Exception $e) {
+							$errorMessages[] = $e->getMessage();
+						}
+						
+					}
+					
+				}
+				
+			}
+			
+		}
+			
+		return $errorMessages;
+		
 	}
 }
 ?>
